Deps: Update androidx fragment to 1.5.5
Release Notes: https://developer.android.com/jetpack/androidx/releases/
fragment#1.5.5

------------------------------------------------------------------------

Note that this update adds the following 'Options Menu' specific
warnings that need to be resolved in order for this project to compile
again due to the 'allWarningsAsErrors' flag.

Per module, the number of warnings and their corresponding classes are:
- 'image-editor' module (7):
  - CropFragment (4)
  - PreviewImageFragment (3)
- 'WordPress' module (24):
  - UnifiedCommentsEditFragment (2)
  - ScanHistoryFragment (2)
  - MediaPickerFragment (2)
  - PageParentFragment (3)
  - ReaderFragment (2)
  - ReaderPostDetailFragment (4)
  - SiteCreationBaseFormFragment (1)
  - StatsFragment (1)
  - StatsListFragment (3)
  - InsightsDetailFragment (1)
  - StatsDetailFragment (1)
  - InsightsManagementFragment (2)

As such, all the above listed modules, classes and their warnings need
to be migrated to the new 'Options Menu' APIs. All that is going to be
done in subsequent commits.

------------------------------------------------------------------------

PS: In the meanwhile, the 'allWarningsAsErrors' flag is going to be
disabled in order to make it possible for the project to compile so that
each screen migration could be tested in isolation.