Deps: Add androidx activity dependency (1.5.1)
This dependency is added in order to access the activity related
'addMenuProvider(...)' API, which is not available prior to the
AndroidX Activity 1.4.0 release, see below:

https://developer.android.com/jetpack/androidx/releases/activity
#1.4.0-alpha01

Currently, and although the project is indeed using AndroidX Activity as
a transitive dependency, those point to older versions of it. Below is a
breakdown per module, and in parenthesis, the version it uses:
- image-editor (1.2.4)
- editor (1.1.0)
- WordPress (1.3.1)

FYI: The activity related 'addMenuProvider(...)' API is needed because
otherwise the 'Options Menu' migration can't commence.

------------------------------------------------------------------------

Release Notes: https://developer.android.com/jetpack/androidx/releases/
activity#1.5.1

------------------------------------------------------------------------

Note that the '1.6.1' update requires libraries and applications that
depend on it to compile against version 33 (Android 13) or later of the
Android APIs. However, this project is currently compiled against
android-31 (Android 12).

Thus, until the 'compileSdkVersion' of this project gets updated to '33'
updating to '1.9.0' isn't possible.

------------------------------------------------------------------------