Analysis: Resolve options menu deprecated warnings on scan history
Warning Messages:
- "'setHasOptionsMenu(Boolean): Unit' is deprecated.
Deprecated in Java"
- "'onOptionsItemSelected(MenuItem): Boolean' is deprecated.
Deprecated in Java"

------------------------------------------------------------------------

Adding 'requireActivity().addMenuProvider(this, viewLifecycleOwner)' and
replacing the 'onOptionsItemSelected' with the 'onMenuItemSelected'
equivalent 'MenuProvider' related interface signatures is now the
recommended way forward.

Also, the empty 'MenuProvider.onCreateMenu(...)' interface
implementation was added since this isn't optional. There is no default
implementation for this interface function. As such, one would need to
implement this interface when 'MenuProvider' is utilized.

------------------------------------------------------------------------

For more info see:
- androidx.fragment.setHasOptionsMenu (Documentation):
https://developer.android.com/reference/androidx/fragment/app/Fragment
#setHasOptionsMenu(boolean)
- androidx.fragment.onOptionsItemSelected (Documentation):
https://developer.android.com/reference/androidx/fragment/app/Fragment
#onOptionsItemSelected(android.view.MenuItem)