create a Result class for the Fetch UseCase
This is needed because detekt complains about TooGenericExceptionCaught in
the ViewModel class. Also, I noticed the real implementation of the Prompts
repository / Store (inside FluxC) doesn't return exceptions anyway, so having
a domain-level result class to propagate the Store results will make more sense
anyway. Right now the Result class was addded inside the UseCase and is not
generic. Not sure if there's value in making this kind of Result class generic
in the codebase, it seems other places are also using custom Result classes.