use the local site id to avoid duplicate keys in LazyColumn
The key in the LazyColumn was coming from FluxC `siteId` field, which contains
the remote ID of the site, but only from wp.com sites. The issue here is that
the `siteId` is 0 for self-hosted sites, meaning that any users with more than
1 self-hosted site would cause duplicate keys in the LazyColumn, crashing the
app.

This commit changes the key to be the local ID (primary key in the local
WellSQL DB inside FluxC), so it cannot have duplicates.