reformat all files after merging trunk into this branch
One thing I noticed is that empty lines were added between annotated properties
this time. After some investigation, it seems Android Studio is unable to do all
formatting at once so before, when I ran reformat once, it just added a line
break after each annotation, and this time it added blank lines between
annotated properties, which is also part of the convention.

With that in mind, I ran reformatting twice, so all changes would be covered on
top of the most recent changes in trunk.