Add a test for migration error logging
The solution used in this commit is to copy the logging wrapper from
fluxc in the unit test source path so that it remains compatible with
the copied AppLog class (which is also copied in a similar manner, as a
work-around). This prevents build errors when the logging wrapper's mock
is verified to have been invoked.