Fix: Make suggestion tokenizer nullable to fix inflate exception
This fix commit tackles this crash: https://github.com/
wordpress-mobile/WordPress-Android/pull/17741#issuecomment-1381943601

By making 'mSuggestionTokenizer' nullable and then guarding its usage
within 'enoughToFilter()' to return false when the tokenizer is null
seems to be enough to solve this inflate related crash.

Note that normally the 'mSuggestionTokenizer' gets initialized via the
'initializeWithPrefix(...)'. However, that usually happens after the
view is being inflated and not before. Thus, this guarantees the crash
if the 'mSuggestionTokenizer' is not being guarded against nullability.