Download the current Latvian translations
 - I ran `bundle exec fastlane download_translations` once, which created a commit that updated `available_languages.xml` and all the `values-*/strings.xml` files, then created a commit.
 - I then amended that commit made by the script to remove all the updated `values-*/` files except `values-lv/`.

This is because this PR will target `develop`, while we should only include string updates during the code freeze phase, so I didn't want to update the other `values-*` translations of other locales at that stage to avoid disrupting the usual release process. But it still felt OK to keep the Latvian ones, as those are entirely new in the repo, and having them will allow to test the Installable Build in Latvian.