Deps: Remove unused androidx vectordrawable animated version
The dependency that was associated with this version was removed as part
of this f2a8b18409c39242227691a2223f0168b4f40786 commit. The version
should have been removed as part of this commit but got slipped from
being removed. As such, this version is unused and can be now safely
removed.