also check for prompts being active in "create site" bottom sheet
We should not display the Blogging Prompts action in the "Create Site" bottom
sheet if the user has disabled the Prompts in the "Site Settings" as well as if
the card would not be visible because of other conditions (ex: was skipped for
today). To do so, we should use the `isPromptsFeatureActive` function to make
sure all conditions are checked.