Fix failing reader tests affected by vararg matchers change
In Mockito 5, the behavior of `any()` for matching vararg parameters has changed. Previously it was matching invocations with 0 to any number or arguments, now it only matches invocations with exactly one argument. As a result, to mock invocations with any number of arguments, we need to use `anyArray<T>()` or `any<Array<*>>()` instead. I went with `anyArray` because it felt less verbose to me.
Source: https://github.com/mockito/mockito/releases/tag/v5.0.0