Tests: Disable ui test auto-syncing with compose in JP app
This is a workaround to support interacting with the compose ui from `LoginPrologueRevampedFragment`, which has an infinite animation. By disabling the auto-sync between tests and compose ui, we avoid `ComposeNotIdleException`.
See: https://developer.android.com/jetpack/compose/testing#disable-autosync

The change was added in the setup of `BaseTest` to support login in JP app tests.