Deps: Remove room dependency
This set of 'Room' related test dependencies, which were only targeting
UI test, were added as part of the below PR and only in order to fix
some connected tests from failing due to classpath errors, see below:

PR: Add room libraries to fix connected tests #14974
https://github.com/wordpress-mobile/WordPress-Android/pull/14974

------------------------------------------------------------------------
e: Supertypes of the following classes cannot be resolved. Please make
sure you have the required dependencies in the classpath:
    class org.wordpress.android.fluxc.persistence.WPAndroidDatabase,
     unresolved supertypes: androidx.room.RoomDatabase
------------------------------------------------------------------------

However, these kind of errors were also appearing on unit tests at that
point. Still, after a while and even though the above fix PR only
targeted Android tests, it appeared that the unit tests were also
working as expected, at least for a while, and at least for most of our
engineers. Matthew Kevins was an exception to that as he reported that
he experienced this issue again 10 months ago. But, even then, by just
rebuilding the project, this issue seems to resolve itself.

As such, with this commit, I recommend trying our removing this 'Room'
dependency from the project and then monitor this change for a few
days/weeks/months to see if this problems reappears. Or else, whether
this problem has been now automatically resolved by itself, due to
maybe an IDE, AGP or any other update that might have potentially fix
this issue automatically for us.

With this change, the project will also become more lightweight as this
will mean maintaining one less dependency overall.

PS: Also, WCAndroid, which also using FluxC and thus utilizing 'Room'
the same way this project does, never applied the same fix, that is, to
add room libraries to fix any connected or unit test. This means that
this extra 'Room' dependency is unnecessary on this project as well and
should/could be potentially removed for good.