Deps: Remove the no longer necessary androidx work gcm dependency
This dependency was added as part of this
82b0c50cb5cb44a320e1b8e6bcf5a7c71e1efeb7 commit (see PR below), which
supported the use of GCMNetworkManager as a scheduler when Google Play
Services is available for API levels <= 22. This was an optional
dependency that helped with more reliable and performant background
processing on older API versions as GCMNetworkManager on api <= 22 (see
2.2.0 release notes below). However, since this project is now on
api >= 24 (see 'minSdkVersion'), this dependency is no longer
necessary and thus can be removed.

- PR: Issue/10572 update work manager #13542
https://github.com/wordpress-mobile/WordPress-Android/pull/13542
- Work Manager 2.2.0: https://developer.android.com/jetpack/androidx/
releases/work#2.2.0

------------------------------------------------------------------------

Also, the 'Dependency Analysis' plugin advised in favor of that removal,
thus, it is most probably safe to do this removal at this point:

Dependencies which should be removed or changed to runtime-only:
  runtimeOnly 'androidx.work:work-gcm:2.7.1' (was implementation)