Deps: Remove media dependency
This 'Media' related dependency was added as part the migration process
to 'AndroidX' for this module back in Jun 2019
(b7ac12d358f7238beb0ca48fc56916ff3c6b4128).

There is no 'androidx.media' usage in this project. As such, this
leftover from back then can be now safely removed.

------------------------------------------------------------------------

FYI: However, there exist a number of transitive dependencies that are
using 'androidx.media', those are the below:
1. com.zendesk:support
2. com.google.android.exoplayer:exoplayer
3. project :libs:editor
4. org.wordpress:login

------------------------------------------------------------------------
1. com.zendesk:support
     com.zendesk:guide
       androidx.legacy:legacy-support-v4
         androidx.media:media
2. com.google.android.exoplayer:exoplayer
     com.google.android.exoplayer:exoplayer-ui
       androidx.media:media
3. project :libs:editor
     org.wordpress:aztec
       androidx.legacy:legacy-support-v4
         androidx.media:media
     org.wordpress-mobile.gutenberg-mobile:react-native-gutenberg-bridge
       com.github.wordpress-mobile:react-native-video
         com.google.android.exoplayer:exoplayer
           com.google.android.exoplayer:exoplayer-ui
             androidx.media:media
         androidx.media:media
4. org.wordpress:login
     androidx.media:media
------------------------------------------------------------------------

From the above 4 dependencies that are using 'androidx.media' via a
transitive dependency:
1. For 'com.zendesk:support', this 'Media' related dependency is part of
the legacy 'AndroidX Support V4' dependency. Thus, I recommend that this
'Media' transitive  dependency is not added explicitly, just because of
those legacy 'AndroidX Support V4' dependencies.
2. For 'com.google.android.exoplayer:exoplayer', this 'Media' related
dependency is closely related to 'Exoplayer UI'. Now, as soon as this is
updated, its 'Media' dependency should be updated as well, most probably
to newer versions of 'Media 2', or even the 'Media 3' libraries. Thus, I
recommend that this 'Media' transitive dependency is not added
explicitly, just because of that 'ExoPlayer' dependency.
3. For 'org.wordpress:aztec'/'org.wordpress-mobile.gutenberg-mobile',
flowing through the ':libs:editor' project dependency, this 'Media'
related dependency of theirs is part of either the legacy
'AndroidX Support V4' dependency or the 'ExoPlayer' player dependency.
As per the two above points, I recommend that this 'Media' transitive
dependency is not added explicitly, just because of those.
4. For 'org.wordpress:login', this 'Media' related dependency was
also added as part the migration process to 'AndroidX'.

As such, and even though the 'Dependency Analysis' plugin recommends to
declare this 'androidx.media' dependency directly (see below), for the
above reasons, this dependency is removed nevertheless.

------------------------------------------------------------------------
Transitively used dependencies that should be declared directly as
indicated:
  ...
  implementation 'androidx.media:media:1.2.1'
  ...
------------------------------------------------------------------------