propagate isPromptsCardEnabled so it keeps its value
When mapping to the Reminders UI model, the `isPromptsCardEnabled` field, which
represents the "Show prompts" setting, was not being propagated properly so when
coverting the UI model back to a Reminders model, that information was lost.

This change makes sure the value is added to the UI model as well, even though
it is not really used visually in that screen. Unfortunately those 2 features
(reminders and prompts) are coupled and use the same FluxC model for their
settings.