point to a setting default PR in FluxC
The FluxC PR caontains the changes for setting a default value for the "show
prompts" setting when the Reminders object does not exist locally. It uses the
`isPotentialBloggingSite` as default in that case, making sure non-blogging
sites have the Prompts feature set to disabled when updating to this version but
are able to enable it from the settings.