Deps: Update google exoplayer to 2.18.1
Release Notes: https://github.com/google/ExoPlayer/releases/tag/r2.18.1

------------------------------------------------------------------------

Note that the '2.18.2' update requires libraries and applications that
depend on it to compile against version 33 (Android 13) or later of the
Android APIs. However, this project is currently compiled against
android-31 (Android 12).

Thus, until the 'compileSdkVersion' of this project gets updated to '33'
updating to '2.18.2' isn't possible.

------------------------------------------------------------------------