Analysis: Suppress recyclerview adapter position deprecation warning
This started appearing because, as part of the 'ExoPlayer' update to
'2.18.1', its transient 'RecyclerView' dependency got updated
as well, from '1.1.0' to '1.2.1', which caused this deprecation warning.

Warning Message: "'getter for adapterPosition: Int' is deprecated.
Deprecated in Java"

Explanation: "This method is confusing when adapters nest other
adapters. If you are calling this in the context of an Adapter, you
probably want to call getBindingAdapterPosition or if you want the
position as RecyclerView sees it, you should call
getAbsoluteAdapterPosition."

This deprecated warning is suppressed, that is, instead of being
resolved, since a resolution is out of the scope of this 'ExoPlayer'
update.

For more info see:
- RecyclerView.ViewHolder#getAdapterPosition (Doc):
https://developer.android.com/reference/androidx/recyclerview/widget/
RecyclerView.ViewHolder#getAdapterPosition()