Deps: Update androidx recyclerview to 1.2.1
Release Notes: https://developer.android.com/jetpack/androidx/releases/
recyclerview#recyclerview-1.2.1

------------------------------------------------------------------------

Note that this update adds the following 'Adapter Position' specific
warnings that need to be resolved in order for this project to compile
again due to the 'allWarningsAsErrors' flag.

Per module, the number of warnings and their corresponding classes are:
- 'image-editor' module (1):
 - PreviewImageViewHolder.kt (1)
- 'WordPress' module (23):
 - TrailingLabelViewHolder.kt (1)
 - UsernameChangerRecyclerViewAdapter.java (1)
 - SitePickerAdapter.java (3)
 - MediaGridAdapter.java (6)
 - PeopleListFragment.java (1)
 - PluginBrowserActivity.java (1)
 - PluginListFragment.java (1)
 - MultiSelectRecyclerViewAdapter.java (1)
 - SiteSettingsTagListActivity.java (1)
 - PublicizeAccountChooserListAdapter.java (1)
 - ReaderBlogAdapter.java (1)
 - ReaderCommentAdapter.java (1)
 - ReaderPostAdapter.java (1)
 - ReaderSiteSearchAdapter.java (1)
 - StockMediaPickerActivity.java (2)

As such, all the above listed modules, classes and their warnings need
to be migrated to the new 'Adapter Position' APIs. All that is going to
be done in subsequent commits.

For more info see:
- Release Notes (1.2.0): https://developer.android.com/jetpack/androidx/
releases/recyclerview#recyclerview-1.2.0
- RecyclerView.ViewHolder#getAdapterPosition (Doc):
https://developer.android.com/reference/androidx/recyclerview/widget/
RecyclerView.ViewHolder#getAdapterPosition()

------------------------------------------------------------------------

Also, as part of the '1.2.0-alpha02' it seems that this and higher
versions of RecyclerView are not compatible with lower versions of
ViewPager2. The recommendation is to update to newer version of
'View Pager 2', the '1.1.0-alpha01' or above.

- Release Notes (1.2.0-alpha02): https://developer.android.com/jetpack/
androidx/releases/recyclerview#recyclerview-1.2.0-alpha02

This seems to means that updating 'Recycler View' to '1.2.1' might mean
that 'View Pager 2' needs to be update to '1.1.0' as well. However, it
seems that the newest version of 'View Pager 2' is the '1.1.0-beta01',
and that has been this way since August 3, 2021. This means that there
is no production ready version of 'View Pager 2' so far, and more so,
that this library might have been abandoned. Also, the fact that the
'1.2.1' release notes don't suggest this update might also mean that
this requirement is no longer applicable, that is, at least on newer,
post alpha releases of 'Recycler View'.

- View Pager 2: https://developer.android.com/jetpack/androidx/releases/
viewpager2