Analysis: Resolve all adapter position deprecated warnings
Warning Messages:
- "'getAdapterPosition()' is deprecated"

Explanation: "This method is confusing when adapters nest other
adapters. If you are calling this in the context of an Adapter, you
probably want to call 'getBindingAdapterPosition' or if you want the
position as 'RecyclerView' sees it, you should call
'getAbsoluteAdapterPosition'."

------------------------------------------------------------------------

If those deprecation warnings were left unresolved, the
'getAdapterPosition()' method will anyway redirect the call to
'getBindingAdapterPosition()'. As such, it is better to be explicit now
and call this newly added 'getBindingAdapterPosition()' API directly,
and only change that to 'getAbsoluteAdapterPosition()' if and whenever
necessary.

------------------------------------------------------------------------

For more info see:
- RecyclerView.ViewHolder#getAdapterPosition (Doc):
https://developer.android.com/reference/androidx/recyclerview/widget/
RecyclerView.ViewHolder#getAdapterPosition()