Add support for passing vendor string to domain suggestions api
Replaces the former `includeVendorDot` boolean from the `SiteStore.SuggestDomainsPayload` request payload, which translated to passing `"vendor": "dot""` to the `/domains/suggestions` REST API with a `vendor` string, thus allowing to pass custom vendor strings if needed.
Since the API supports adding new `vendor` strings and customizing the response based on the value, it felt naturally to reflect this in the API client.

The former constructor accepting `includeVendorDot` remains in code to ensure backwards compatibility but it's deprecated and recommended to update the clients to use the primary payload constructor. Autocompletion support has been added for migrating to the new specs which supports extended functionality, including the previously available one.