always postState in Blaze and Prompts sources
Not posting a state and just updating the `refresh` status doesn't cause the
MySiteTabFragment to check if the loading indicator needs to be hidden, as that
is only checked when a UI State is updated. This fix causes all updates to be
propagated and the check to be called for Blaze and Prompts card sources.