update icon even when it's the same
since we return "Default" when the icon ID has not been found, it's possible
(during development) that you have an old icon ID set in the Prefs but it
is converted to "Default". In that scenario, you wouldn't be able to set the
"real" default as an icon because of that, so easier to just let it be saved.

If in a normal scenario the user saves the same icon again, nothing really
happens (the app doesn't close or anything).