Deps: Add retrofit transitively used dependencies on wordpress
It is generally recommended that transitively used dependencies should
be declared directly.

------------------------------------------------------------------------

Actually, updating 'Zendesk', which was the dependency that was
providing this transitively used 'Retrofit' dependency to the project,
was causing a compilation failure. This is due to the fact that
the [TenorGifClient] is directly using this dependency, but could no
longer resolve it.

PS: This is (most probably) due to the fact that this 'Retrofit'
dependency was no longer available on newer 'Zendesk' versions.