Deps: Add google flexbox transitively used dependencies on wordpress
It is generally recommended that transitively used dependencies should
be declared directly.

------------------------------------------------------------------------

Actually, updating 'Zendesk', which was the dependency that was
providing this transitively used 'Flexbox' dependency to the project,
was causing a compilation failure. This is due to the fact that the
[BloggingPromptsOnboardingDialogFragment],
[BloggingPromptCardViewHolder] and [WPEditTextWithChipsOutlined] classes
are directly using this dependency, but could no longer resolve it.

PS: This is (most probably) due to the fact that this 'Flexbox'
dependency was no longer available on newer 'Zendesk' versions.