Deps: Remove dagger exclusion from zendesk dependency
This 'exclude' related configuration was added back in Jun 13, 2018
while migrating 'Zendesk' to v2 of 'Zendesk' SDK, or else replacing
'Helpshift' with 'Zendesk' (a13b64606e2f26cd18bcffaa8066fda17f24071f).

Back then, it was probably necessary to exclude 'Dagger' to have the
project compile (due to duplicate classpath error or otherwise).

However, it seems that it is no longer necessary to have such a
configuration where 'Dagger' is excluded. As such, this change is
reverted. As long as the project compiles and 'Zendesk' works as
expected, this change can be consider safe.