force White content color in StaticPoster button
It uses the PrimaryButton, but since it is defining custom colors, it ends up
losing the ability to set the correct color, since the current `onPrimary` color
for Dark mode is Black.

Not sure if that `onPrimary` color is correct, as primary color is the same for
Light and Dark modes, but this fix is less risky as it just touches the part we
know is wrong.