refresh appConfigs upon OnAuthenticationChanged
This should always be done if the auth changed (login or logout) as one of
the rules the Remote Configs support is: is_user_not_authenticated, which means
that some FeatureConfigs return different values depending on the authentication
status.

This was causing several issues, since the jp_removal_self_hosted was one of
those remote fields that was returning true when the app opened without a logged
in user, so even after login it was using the stale value of `true` causing:
- Dashboard and Bottom Nav to not show up for WP.com accounts
- Jetpack Plugin overlay to show for WP.com accounts

None of those should be happening at the current jetpack feature removal phase
for WP.com users.