Analysis: Narrow down ignore config for icon dip size
The 'IconDipSize' error is only applicable for the 'drawable-hdpi'
related resource and not for the rest 2 of them, the 'drawable-xhdpi'
and 'drawable-xxhdpi' (there are 3 such resources in total). As such, it
is better to be explicit on that to narrow down the specific error that
is being ignored.

FYI: The default severity of 'IconDipSize' is warning.
Link: http://tools.android.com/tips/lint-checks