Analysis: Narrow down error config for unused resources
The 'UnusedResources' error for the 'editor' module, is only applicable
for the 'drawable', 'drawable-hdpi', 'layout/alert_create_link.xml',
'layout/alert_image_options.xml', 'values/colors.xml',
'values/wp_colors.xml', 'values/dimens.xml' and 'values/styles.xml'
related resources and not for the rest of them. As such, it is better to
be explicit on that to narrow down the specific error that is being
ignored.

Also, for some of those, 'UnusedResources' was suppressed closer to res
source, that is, instead of it being on the 'lint.xml' configuration
file, and that, in order to help fixing those errors easier later on.

FYI: The default severity of 'UnusedResources' is warning.
Link: http://tools.android.com/tips/lint-checks