Build: Enable all warnings as errors lint options flag
With that change all Lint warnings are now treated as errors. As such,
it is no longer necessary to have an explicit warnings-as-errors
configuration within the 'lint.xml' file. Now, the only necessary
configuration within this 'lint.xml' file is related to specific errors
as a whole (like 'MissingTranslation', 'ExtraTranslation' and
'PluralsCandidate') or ignoring a set of errors based on a path or a
regular expression (like 'IconDipSize', 'UnusedResources' and 'Typos').

FYI: The current total amount of errors are about 500.
Link: https://developer.android.com/reference/tools/gradle-api/
4.1/com/android/build/api/dsl/LintOptions