Build: Add lint baseline grouped by build and source set categories
This Lint baseline suppresses the existing warnings-as-errors issues
(about 500 in total) to have Lint running as if the codebase has zero
issues, but still failing the build on new warnings, just like it used
to fail the build on new errors. This helps to start utilising the new
warnings-as-errors Lint configuration immediately without having to
worry about any existing warnings.

Any new warnings that will be added to the codebase right after will be
reported and the build will fail accordingly.

FYI: Actual number of current warnings per app/variant:
- WordPress/Debug:   486
- WordPress/Release: 481
- Jetpack/Debug:     492
- Jetpack/Release:   487

PS: I excluded the further breakdown of Debug to the Wasabi and Jalapeno
variant as the only difference there is the extra generated
'MissingFirebaseInstanceTokenRefresh' issue, which is variant specific.