fix remote field repository to properly add defaults
Field defaults were ONLY being added when the local remote field DB table was
empty, but in cases of app update, that table will already have some values and
any new remote configs were not being added to it.

This fix makes sure to add missing fields if none was found in the currently
existing table.