fix: Enable unit tests to mock URL
The Android-specific Uri class for encoding query parameters doesn't work with unit tests.

As such, it's been replaced with the java.net.URLEncoder class with this commit.

As the URLEncoder class isn't supported on versions of Android lower than 13 (and the app still officially supports versions down to Android 7), conditional logic is used to incorporate both.