Analysis: Resolve manifest order lint warning for debug
Warning Message: "'<uses-permission>' tag appears after
'<application>' tag"

Explanation: "The <application> tag should appear after the elements
which declare which version you need, which features you need, which
libraries you need, and so on. In the past there have been subtle bugs
(such as themes not getting applied correctly) when the <application>
tag appears before some of these other elements, so it's best to order
your manifest in the logical dependency order."

------------------------------------------------------------------------

Reordering the '<application>' and moving it after the
'<uses-permission>' tags fixes this warning.