Analysis: Suppress exported content provider lint warning for wordpress
Warning Message: "Exported content providers can provide access to
potentially sensitive data"

Explanation: "Content providers are exported by default and any
application on the system can potentially use them to read and write
data. If the content provider provides access to sensitive data, it
should be protected by specifying export=false in the manifest or by
protecting it with a permission that can be granted to other
applications."

------------------------------------------------------------------------

This exported content provider was added as part of
dc7ba567078f2f4a470554c72382bf6bd9029276 in order to enable content
migration from WordPress to Jetpack. The migration is currently still
in-progress (ish) so suppressing this for the the time being and at
least until the migration is fully complete.