Analysis: Suppress missing application icon lint warning for jetpack
Warning Message: "You should set an icon for the application as whole
because there is no default. This attribute must be set as a reference
to a drawable resource containing the image
(for example @drawable/icon)."

------------------------------------------------------------------------

Suppress this warning for the 'jetpack' source set should be enough for
the time being.