remove NestedScrollView and use NestedWebView
The NestedScrollView behavior paired with the WebView was causing the webview
height to increase indefinitely while the tumblr login was trying to center its
content vertically, causing the login form to be pushed down.

This likely didn't happen with other services because they do not center their
content vertically or at least don't do it the same way tumblr was doing.

The correct solution though was to use the NestedWebView, which was created to
avoid these kind of scroll issues that can affect WebViews inside scrolling
layouts.