Deps: Remove duplicate compose lifecycle version
Both 'androidxComposeLifecycleVersion' and 'androidxLifecycleVersion'
point to '2.5.1' version. As such, using the compose related
'androidxComposeLifecycleVersion' is not necessary and might actually
cause problems if one is updated and not the other.

FYI: It seems that the 'AndroidX Lifecycle' library, both for
'lifecycle-*' and 'lifecycle-viewmodel-compose' artifacts, has started
pointing towards the same version. The 'lifecycle-viewmodel-compose'
artifact was previously pointing to a different version, but that has
changed since. As such it is better to align the versioning and use a
similar approach when updating this library to avoid incompatibility or
other such issues.