Analysis: Suppress unused material scaffold padding parameter lint warns
Warning Message: "Content padding parameter it is not used"

Explanation: "The content lambda in Scaffold has a padding parameter
which will include any inner padding for the content due to app bars.
If this parameter is ignored, then content may be obscured by the app
bars resulting in visual issues or elements that can't be interacted
with."

------------------------------------------------------------------------

These Lint warnings are suppressed, that is, instead of them being
resolved, since a resolution would require a proper investigation and
testing. As such, it might be best to ignore this as out of scope, for
now, and so as to not introduce any breaking changes to this scaffold
related functionality overall.