Analysis: Suppress freq. changed state read in composition lint warns
Warning Message: "Frequently changing state should not be directly read
in composable function"

Explanation: "This property is observable and is updated after every
scroll or remeasure. If you use it in the composable function directly,
it will be recomposed on every change, causing potential performance
issues including infinity recomposition loops. Prefer wrapping it with
derivedStateOf to use calculation based on this property in composition
or collect changes inside LaunchedEffect instead."

------------------------------------------------------------------------

These Lint warnings are suppressed, that is, instead of them being
resolved, since a resolution would require a proper investigation and
testing. As such, it might be best to ignore this as out of scope, for
now, and so as to not introduce any breaking changes to this scroll
related functionality overall.