Fix: Remove unsupported placeholders and errors from coil's async image
Unfortunately, Compose's 'Painter' and 'PainterResources' APIs do not
support '<layer-list/>', or anything other than a 'VectorDrawable' or a
rasterized asset type (PNG, JPG, etc).

As such, when one tried to do use '<layer-list/>' the below exception is
thrown:

------------------------------------------------------------------------
E  FATAL EXCEPTION: main
   Process: com.jetpack.android.beta, PID: 14827
   java.lang.IllegalArgumentException: Only VectorDrawables and
    rasterized asset types are supported ex. PNG, JPG
   	 at androidx.compose.ui.res.PainterResources_androidKt
   	  .loadVectorResource(PainterResources.android.kt:94)
   	 at androidx.compose.ui.res.PainterResources_androidKt
   	 .painterResource(PainterResources.android.kt:65)
------------------------------------------------------------------------

As such, in order to support placeholders and errors for 'AsyncImage',
both, the 'bg_rectangle_placeholder_globe_margin_8dp' and
'bg_rectangle_placeholder_globe_32dp' drawables, need to be converted to
a 'VectorDrawable' or a rasterized asset type.

In the meanwhile, those placeholders and errors are removed in order to
progress with this Coil update. Later on, with some design help, both,
the placeholders and errors can get added back.

------------------------------------------------------------------------

PS: I also tried to convert the '<layer-list/>' to a Compose image, but
this is not working with 'Painter' and 'PainterResources', which is the
API that 'AsyncImage' uses, see below:

Image(
    modifier = Modifier
        .padding(8.dp) // For bg_rectangle_placeholder_globe_margin_8dp
        .clip(shape = RectangleShape)
        .background(color = colorResource(id = R.color.placeholder)),
    painter = painterResource(id = R.drawable.ic_globe_white_24dp),
    contentDescription = "...",
)