Analysis: Resolve companion object of enum class uninitialized warning
Warning Message: "Companion object of enum class 'FieldType' is
uninitialized here. This warning will become an error in future
releases. See https://youtrack.jetbrains.com/issue/KT-49110 for details"

FYI: This warning will be raised to error in Kotlin version '1.9.0'.

------------------------------------------------------------------------

Existing Behavior: "When one accesses companion object or its members of
an enum from the enum entry initializer, they can get an NPE there
because the entries of an enum are initialized before the companion
object of that enum at runtime.

Some of such code was prohibited in Kotlin 1.1 (see KT-11769), but there
are still ways to access companion from enum initializer.

Note that sometimes the code does not throw an exception, just like this
case, however it still observes an uninitialized state of a companion
property, and thus needs fixing to avoid any future misbehavior."

Proposed Changes: "It's proposed to prohibit all kinds of access to the
companion object of an enum from an enum entry initializer."

Impact: "Code like this won't compile. This is a rare case, because in
most cases such code throws NPE or acts differently than users expect,
so it should be fixed right away."

------------------------------------------------------------------------

This commit resolves this warning by creating a regular object inside
the enum itself. Initialization of nested objects is not bound to
initialization of the enum itself (in opposite to companion objects),
so it's safe to access them during initialization of enum entries.

Dmitriy Novozhilov (JetBrains): https://youtrack.jetbrains.com/issue/
KT-49110/Prohibit-access-to-members-of-companion-of-enum-class-from-
initializers-of-entries-of-this-enum#focus=Comments-27-7246774.0-0

PS: Another idea would be to extract these function to a file, but I
chose the 'regular object' idea so as to include only the bare minimum
changes here.